moved the url parsing and navigation logic to QuranEngine
